Restructuring chats
