+++ src/config/locales/en-US.json
@@ -1282,5 +1282,40 @@
     "btn-dismiss": "Dismiss",
     "btn-later": "Remind Later",
     "btn-ignore": "Forever"
+  },
+  "chats": {
+    "anonymous": "Unknown user",
+    "new_messages": "New messages",
+    "channels": "Channels",
+    "join": "Join",
+    "users": "Users",
+    "connected": "Connected",
+    "connecting": "Connecting...",
+    "editing_message": "Editing message",
+    "people_joined": "{count} people joined",
+    "remove_message": "Remove message?",
+    "remove": "Remove",
+    "remove_message_body": "This will remove the message for everyone in the channel.",
+    "header_with_members": "{title} - {count} members",
+    "unfavorite": "Remove favorite",
+    "favorite": "Favorite",
+    "unmute": "Unmute",
+    "mute": "Mute",
+    "leave": "Leave channel",
+    "search_placeholder": "Search channels and users",
+    "login_required": "Sign in to start chatting with your communities.",
+    "unable_to_leave_channel": "Unable to leave channel. Please try again.",
+    "unable_to_start_dm": "Unable to start chat. Please try again.",
+    "unable_to_join_channel": "Unable to join channel. Please try again.",
+    "unable_to_mark_read": "Unable to mark channel as read. Please try again.",
+    "message_placeholder": "Message",
+    "banned_from_chat": "Unusual activity detected. Please try again after some time.",
+    "channel_options": "Channel options",
+    "empty": "No chats yet. Join a community to see its channels here.",
+    "mark_as_read": "Mark as read",
+    "no_channels": "No channels yet. Search to join or start a new conversation.",
+    "no_more_messages": "No more messages here",
+    "reaction_error": "Failed to update reaction",
+    "thread_empty": "No messages yet. Say hello to get the conversation started."
   }
 }
