+++ src/screens/chats/children/MessageReactions.tsx
@@ -1,0 +1,69 @@
+import React from 'react';
+import { View, Text } from 'react-native';
+import { getEmojiDisplay } from '../utils/messageFormatters';
+import { chatThreadStyles as styles } from '../styles/chatThread.styles';
+
+interface ChatReaction {
+  emoji_name: string;
+  user_id: string;
+  create_at?: number;
+}
+
+interface MessageReactionsProps {
+  reactions: ChatReaction[] | undefined;
+  isOwnMessage: boolean;
+  bootstrapUserId: string;
+}
+
+export const MessageReactions: React.FC<MessageReactionsProps> = React.memo(
+  ({ reactions, isOwnMessage, bootstrapUserId }) => {
+    if (!reactions || reactions.length === 0) {
+      return null;
+    }
+
+    // Group reactions by emoji_name
+    const groupedReactions: Record<string, ChatReaction[]> = {};
+    reactions.forEach((reaction) => {
+      if (!groupedReactions[reaction.emoji_name]) {
+        groupedReactions[reaction.emoji_name] = [];
+      }
+      groupedReactions[reaction.emoji_name].push(reaction);
+    });
+
+    return (
+      <View
+        style={[
+          styles.reactionsContainer,
+          isOwnMessage ? styles.reactionsContainerOwn : styles.reactionsContainerOther,
+        ]}
+      >
+        {Object.entries(groupedReactions).map(([emojiName, reactionList]) => {
+          const emojiDisplay = getEmojiDisplay(emojiName);
+          const count = reactionList.length;
+          const hasCurrentUserReaction = reactionList.some((r) => r.user_id === bootstrapUserId);
+
+          return (
+            <View
+              key={emojiName}
+              style={[styles.reactionPill, hasCurrentUserReaction && styles.reactionPillActive]}
+            >
+              <Text style={styles.reactionEmoji}>{emojiDisplay}</Text>
+              {count > 1 && (
+                <Text
+                  style={[
+                    styles.reactionCount,
+                    hasCurrentUserReaction && styles.reactionCountActive,
+                  ]}
+                >
+                  {count}
+                </Text>
+              )}
+            </View>
+          );
+        })}
+      </View>
+    );
+  },
+);
+
+MessageReactions.displayName = 'MessageReactions';
