+++ src/screens/chats/styles/chats.styles.ts
@@ -1,0 +1,251 @@
+import EStyleSheet from 'react-native-extended-stylesheet';
+
+export const chatsStyles = EStyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '$primaryBackgroundColor',
+  },
+  content: {
+    flex: 1,
+  },
+  header: {
+    paddingHorizontal: 16,
+    backgroundColor: '$primaryBackgroundColor',
+    alignItems: 'flex-end',
+  },
+  title: {
+    fontSize: 22,
+    fontWeight: '700',
+    color: '$primaryBlack',
+  },
+  subtitle: {
+    marginTop: 4,
+    color: '$primaryDarkText',
+    fontSize: 14,
+    lineHeight: 20,
+  },
+  statusPill: {
+    position: 'absolute',
+    flexDirection: 'row',
+    alignItems: 'center',
+    backgroundColor: '$primaryLightBackground',
+    paddingHorizontal: 12,
+    paddingVertical: 8,
+    borderRadius: 20,
+    shadowColor: '$shadowColor',
+    shadowOffset: { width: 0, height: 2 },
+    shadowOpacity: 0.1,
+    shadowRadius: 4,
+    elevation: 3,
+    right: 16,
+    bottom: 8,
+  },
+  statusDot: {
+    width: 8,
+    height: 8,
+    borderRadius: 4,
+    marginRight: 6,
+  },
+  statusPillText: {
+    fontSize: 12,
+    color: '$primaryDarkText',
+    fontWeight: '500',
+  },
+  card: {
+    marginTop: 12,
+    padding: 12,
+    backgroundColor: '$primaryLightBackground',
+    borderRadius: 12,
+  },
+  cardTitle: {
+    fontSize: 16,
+    fontWeight: '600',
+    color: '$primaryBlack',
+  },
+  cardSubtitle: {
+    marginTop: 4,
+    color: '$primaryDarkText',
+    fontSize: 13,
+    lineHeight: 18,
+  },
+  channelRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingVertical: 14,
+  },
+  channelAvatar: {
+    width: 40,
+    height: 40,
+    borderRadius: 20,
+    marginRight: 12,
+    borderWidth: 0,
+  },
+  channelAvatarFallback: {
+    width: 40,
+    height: 40,
+    borderRadius: 20,
+    marginRight: 12,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: '$primaryLightBackground',
+  },
+  channelAvatarText: {
+    fontWeight: '700',
+    color: '$primaryDarkText',
+  },
+  channelRowContent: {
+    flex: 1,
+  },
+  channelTitleRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    flexWrap: 'nowrap',
+  },
+  channelName: {
+    fontSize: 15,
+    fontWeight: '600',
+    color: '$primaryBlack',
+    flexShrink: 1,
+  },
+  channelMetaIcon: {
+    marginLeft: 6,
+  },
+  channelMeta: {
+    marginTop: 4,
+    color: '$primaryDarkText',
+    fontSize: 13,
+  },
+  errorText: {
+    color: '$primaryRed',
+    marginTop: 8,
+  },
+  searchContainer: {
+    paddingVertical: 12,
+    borderRadius: 12,
+  },
+  searchInputRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginHorizontal: 16,
+  },
+  searchInputWrapper: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    borderRadius: 16,
+    height: 44,
+    flex: 1,
+    backgroundColor: '$primaryLightBackground',
+    paddingHorizontal: 12,
+  },
+  sortButton: {
+    width: 44,
+    height: 44,
+    borderRadius: 16,
+    backgroundColor: '$primaryLightBackground',
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginLeft: 8,
+    borderWidth: 0,
+  },
+  sortButtonActive: {
+    borderWidth: 2,
+    borderColor: '$primaryBlue',
+  },
+  searchIcon: {
+    marginRight: 8,
+  },
+  searchInput: {
+    flex: 1,
+    paddingVertical: 10,
+    color: '$primaryDarkText',
+  },
+  clearButton: {
+    padding: 4,
+    marginLeft: 8,
+  },
+  searchResults: {},
+  searchSection: {
+    marginTop: 10,
+    marginBottom: 12,
+  },
+  searchRow: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingVertical: 10,
+  },
+  searchRowAvatar: {
+    width: 40,
+    height: 40,
+    borderRadius: 20,
+    marginRight: 12,
+  },
+  searchRowAvatarFallback: {
+    width: 40,
+    height: 40,
+    borderRadius: 20,
+    marginRight: 12,
+    alignItems: 'center',
+    justifyContent: 'center',
+    backgroundColor: '$primaryLightBackground',
+  },
+  searchRowAvatarText: {
+    fontWeight: '700',
+    color: '$primaryDarkText',
+  },
+  searchRowContent: {
+    flex: 1,
+  },
+  searchAction: {
+    paddingVertical: 8,
+    paddingHorizontal: 12,
+    borderRadius: 8,
+    backgroundColor: '$primaryBlue',
+  },
+  searchActionLabel: {
+    color: '$pureWhite',
+    fontWeight: '600',
+  },
+  searchMessageAction: {
+    paddingVertical: 8,
+    paddingHorizontal: 12,
+  },
+  searchMessageActionLabel: {
+    color: '#357ce6',
+    fontWeight: '600',
+  },
+  searchActionIcon: {
+    padding: 8,
+  },
+  sectionHeading: {
+    color: '$primaryDarkText',
+    marginBottom: 6,
+    fontSize: 20,
+  },
+  channelOptions: {
+    paddingHorizontal: 8,
+    paddingVertical: 4,
+  },
+  optionsLabel: {
+    fontSize: 18,
+    color: '$primaryDarkText',
+  },
+  unreadBadge: {
+    minWidth: 26,
+    paddingHorizontal: 8,
+    paddingVertical: 4,
+    borderRadius: 14,
+    backgroundColor: '$primaryBlue',
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginRight: 6,
+  },
+  unreadBadgeText: {
+    color: '$pureWhite',
+    fontWeight: '700',
+    fontSize: 12,
+  },
+  emptyState: {
+    marginTop: 16,
+    color: '$primaryDarkText',
+  },
+});
